requirements.txt
gevent==1.4.0
Werkzeug==0.15.2
Flask==1.0.2
pillow==6.0.0
torch==1.4.0
torchvision==0.5.0